Refactor classes using folders
