+++ Service/src/main/java/SendingTheTelephone/TicketSending.java
@@ -1,4 +1,6 @@
-package org.example;
+package SendingTheTelephone;
+
+import TIcket.Ticket;
 
 public abstract class TicketSending {
     public void shared(Ticket ticket){};
