+++ Service/src/main/java/Users/User.java
@@ -1,4 +1,4 @@
-package org.example;
+package Users;
 
 public class User {
     public void PrintRole(){};
